Some code review fixes
