+++ ui/src/components/LoginForm/SessionTable.module.css
@@ -17,7 +17,7 @@
   cursor: pointer;
 }
 
-.row_now {
+.row[data-highlight] {
   background-color: #eeff9c;
   outline-offset: -2px;
 }
