+++ ui/vite.config.js
@@ -14,6 +14,7 @@ export default defineConfig({
       '/mxcube/api': 'http://127.0.0.1:8081',
       '/socket.io/': { target: 'ws://127.0.0.1:8081', ws: true },
     },
+    allowedHosts : ["lalex.esrf.fr"]
   },
   build: {
     outDir: 'build',
